Enhance NativePHP integration and UI improvements
Added NativePHP alert confirmation for product deletion and improved window/menu setup in NativeAppServiceProvider. Updated nativephp config with new version, author, description, and website. Added and updated application icons. Removed redundant JS confirm from product delete button and cleaned up PDF report link. Minor image update included.

+++ app/Livewire/Tabla/Productos.php
@@ -9,6 +9,7 @@ use App\Models\{
 };
 use Livewire\WithPagination;
 use Livewire\Attributes\Url;
+use Native\Desktop\Facades\Alert;
 
 class Productos extends Component
 {
@@ -36,9 +37,19 @@ class Productos extends Component
     }
 
     public function eliminar(Producto $producto)
-    {
-        $producto->delete();
-    }
+    {   
+       $respuesta =  Alert::new()
+        ->type('warning')
+        ->title('Eliminar')
+        ->buttons(['Confirmar', 'Cancelar'])
+        ->show('¿Seguro que deseas eliminar el producto: '.$producto->nombre_producto.'?');  
+        
+        if($respuesta = 0){
+            $producto->delete();           
+        }
+    }        
+        
+    
 
     public function setSortBy($sortBy){
         if($sortBy === 'NoFiltro') {
